рефакторинг, вынес метод convert() из MovieQuizViewController
