+++ MovieQuiz/Presentation/MovieQuizPresenter.swift
@@ -1,0 +1,35 @@
+//
+//  MovieQuizPresenter.swift
+//  MovieQuiz
+//
+//  Created by Роман Ивановский on 29.07.2023.
+//
+
+import Foundation
+import UIKit
+
+final class MovieQuizPresenter {
+     let questionsAmount: Int = 10
+     var currentQuestionIndex: Int = 0
+    
+    func isLastQuestion() -> Bool {
+        currentQuestionIndex == questionsAmount - 1
+    }
+    
+    func resetQuestionIndex() {
+        currentQuestionIndex = 0
+    }
+    
+    func switchToNextQuestion() {
+        currentQuestionIndex += 1
+    }
+    
+    func convert(model: QuizQuestion) -> QuizStepViewModel {
+        let questionStep = QuizStepViewModel(
+            image: UIImage(data: model.image) ?? UIImage(),
+            question: model.text,
+            questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount)")
+        return questionStep
+    }
+    
+}
